Fix generated headers
The currently generated C file for custom fonts is missing a block at
the top to set LV_LVGL_H_INCLUDE_SIMPLE conditionally as is done in the
current image converter. This will fail to compile depending on how your
include paths are set up as it does for me.

Fixes lvgl/lv_font_conv#116